requirements.txt
azure-storage-blob==12.22.0
psycopg2-binary==2.9.9
mysql-connector-python==9.0.0
typer==0.12.3

